refactor(spindrift): make the registry the only list of commands
The command layer declared each command twice: once in its own module, once
in a re-export barrel, and once in the registry. The barrel existed only so
the registry could assert the two lists agreed.

Delete src/commands/index.ts and the MissingFromRegistry / ExtraInRegistry
assertions it fed. The guarantee that has teeth — no dispatchable name the
command layer does not back — is the registry object's satisfies clause, which
is unchanged. Its mirror was never enforced by the pair of lists either:
create-app.ts has always been a command with no registry entry and tripped
nothing.

Tests import the command they exercise from its own module.
