+++ apps/spindrift/test/commands/job-runs.test.ts
@@ -12,12 +12,10 @@
  * schedule started and nothing here asked for.
  */
 import { describe, expect, test } from 'bun:test';
+import { getAppWorkspace } from '../../src/commands/apps/workspace.ts';
+import { createComponent } from '../../src/commands/components/create.ts';
+import { runComponent } from '../../src/commands/components/run.ts';
 import { createApp } from '../../src/commands/create-app.ts';
-import {
-  createComponent,
-  getAppWorkspace,
-  runComponent,
-} from '../../src/commands/index.ts';
 import type {
   AdapterRegistry,
   Clock,
