+++ apps/spindrift/test/commands/workspace-build-route.test.ts
@@ -15,8 +15,9 @@
 import { describe, expect, test } from 'bun:test';
 import { eq } from 'drizzle-orm';
 import { setAppBuildRoute } from '../../src/commands/apps/build-route.ts';
+import { getAppWorkspace } from '../../src/commands/apps/workspace.ts';
+import { createComponent } from '../../src/commands/components/create.ts';
 import { createApp } from '../../src/commands/create-app.ts';
-import { createComponent, getAppWorkspace } from '../../src/commands/index.ts';
 import type {
   AdapterRegistry,
   Clock,
