+++ apps/spindrift/test/web/app-list-identity.test.tsx
@@ -24,10 +24,11 @@
  */
 import { describe, expect, test } from 'bun:test';
 import { isValidElement, type ReactElement, type ReactNode } from 'react';
+import { deleteApp } from '../../src/commands/apps/delete.ts';
 import { listApps } from '../../src/commands/apps/list.ts';
 import { getAppWorkspace } from '../../src/commands/apps/workspace.ts';
+import { createComponent } from '../../src/commands/components/create.ts';
 import { createApp } from '../../src/commands/create-app.ts';
-import { createComponent, deleteApp } from '../../src/commands/index.ts';
 import type {
   AdapterRegistry,
   Clock,
